MNT: move results and related helpers into its own submodule
